requirements.txt
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
python-dotenv>=1.0.0
litellm>=1.0.0
vectorize-client>=1.0.0
requests>=2.31.0
flask>=2.3.0
mcp[cli]>=1.0.0
httpx>=0.24.0 